requirements.txt
appdirs==1.4.4
asgiref==3.5.2
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.3
django-allauth==0.51.0
django-auto-logout==0.5.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-auth==0.9.5
djangorestframework==3.14.0
fake-useragent==1.1.0
gunicorn==20.1.0
idna==3.4
install==1.3.5
Markdown==3.4.1
oauthlib==3.2.2
parse==1.19.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
pyee==8.2.2
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
w3lib==2.0.1
websockets==10.4
whitenoise==6.2.0
zipp==3.11.0
